+++ packages/pangea-sdk/pangea/asyncio/services/authn.py
@@ -6,6 +6,7 @@
 
 from __future__ import annotations
 
+from collections.abc import Mapping
 from typing import Dict, List, Literal, Optional, Union
 
 import pangea.services.authn.models as m
@@ -479,7 +480,7 @@ class AuthNAsync(ServiceBaseAsync):
         async def create(
             self,
             email: str,
-            profile: m.Profile,
+            profile: Mapping[str, str],
             *,
             username: str | None = None,
         ) -> PangeaResponse[m.UserCreateResult]:
@@ -870,7 +871,7 @@ class AuthNAsync(ServiceBaseAsync):
 
             async def update(
                 self,
-                profile: m.Profile,
+                profile: Mapping[str, str],
                 id: str | None = None,
                 email: str | None = None,
                 *,