+++ packages/pangea-sdk/tests/unit/test_service_base.py
@@ -1,20 +1,52 @@
-import unittest
+from __future__ import annotations
 
+from collections.abc import AsyncIterator, Iterator
+
+import pytest
+
+from pangea.asyncio.services.base import ServiceBaseAsync
 from pangea.services.base import PangeaConfig, ServiceBase
 
 
-class TestServiceBase(unittest.TestCase):
-    def test_service_base(self):
-        token = "token"
-        base = ServiceBase(token, PangeaConfig("domain"))
-        self.assertEqual(token, base.token)
+@pytest.fixture(scope="session")
+def base(request: pytest.FixtureRequest) -> Iterator[ServiceBase]:
+    yield ServiceBase("my_token", PangeaConfig(domain="my_domain"))
+
+
+@pytest.fixture(scope="session")
+async def async_base(request: pytest.FixtureRequest) -> AsyncIterator[ServiceBaseAsync]:
+    yield ServiceBaseAsync("my_token", PangeaConfig(domain="my_domain"))
+
+
+class TestServiceBase:
+    def test_service_base(self, base: ServiceBase) -> None:
+        assert base.token == "my_token"
 
         base.token = "newtoken"
-        self.assertEqual("newtoken", base.token)
+        assert base.token == "newtoken"
 
-    def test_service_base_no_token(self):
-        def no_token():
-            base = ServiceBase(None, PangeaConfig("domain"))
+    def test_service_base_no_token(self) -> None:
+        with pytest.raises(Exception):
+            ServiceBase(None, PangeaConfig(domain="domain"))  # type: ignore[arg-type]
 
-        # This should fail because there is no signed configured
-        self.assertRaises(Exception, no_token)
+    def test_extra_headers(self, base: ServiceBase) -> None:
+        base.request.set_extra_headers({"Host": "foobar"})
+        assert "Host" in base.request._headers()
+        assert base.request._headers()["Host"] == "foobar"
+
+
+class TestServiceBaseAsync:
+    def test_service_base(self, async_base: ServiceBaseAsync) -> None:
+        assert async_base.token == "my_token"
+
+        async_base.token = "newtoken"
+        assert async_base.token == "newtoken"
+
+    def test_service_base_no_token(self) -> None:
+        with pytest.raises(Exception):
+            ServiceBaseAsync(None, PangeaConfig(domain="domain"))  # type: ignore[arg-type]
+
+    def test_extra_headers(self, async_base: ServiceBaseAsync) -> None:
+        async_base.request.set_extra_headers({"Host": "foobar"})
+        assert "Host" in async_base.request._headers()
+        assert async_base.request._headers()["Host"] == "foobar"